Add new tests.
test3: Get a non-existent key.

test4: Add an element to a set, print the number of elements in the
set.

test5: Same as test4, but written without "do" notation. An exercise
in wrapping/unwrapping monads.

test6: A race-free test-and-set macro operation. Good for creating
monotonically increasing integer IDs from strings.